requirements.txt
pygame==2.6.0
python-chess==1.999
stockfish==3.28.0